Use Promise.all() to return with first rejection
Prior to this commit: in the chain of promises created with .reduce(),
the temporally first rejection would usually not be associated with the
first promise in the chain of promises. Therefore when other promises
rejected, the incorrect rejection would usually be passed on to the
following .catch() statement.

Use of Promise.all() is idiomatic and resolves this issue.