incorporate worker shim into library
originally the end user was required to provide their own worker-side
interface to the parallel processing system.

now the that interface is a part of the swirlnet-solver-async library
itself, making things easier.

to make this change, the user now supplies the name of the file that
gets loaded by the worker shim, rather than the name of the worker shim
itself. as a result, "testFile" replaces "workerPath" as an option.